+++ test/parse/recognizer/recognizer_config_spec.ts
@@ -1,10 +1,13 @@
 import { Parser } from "../../../src/parse/parser_public"
+import { createToken } from "../../../src/scan/tokens_public"
 
 describe("The Recognizer's Configuration", () => {
     it("default config values - empty config", () => {
+        const A = createToken({ name: "A" })
+
         class EmptyConfigParser extends Parser {
             constructor() {
-                super([], {})
+                super([A], {})
             }
         }
 
@@ -14,9 +17,11 @@ describe("The Recognizer's Configuration", () => {
     })
 
     it("default config values - no config", () => {
+        const A = createToken({ name: "A" })
+
         class NoConfigParser extends Parser {
             constructor() {
-                super([])
+                super([A])
             }
         }